KP-715 Add "Overwrite Existing" parameter (#6)

+++ app.js
@@ -1,12 +1,22 @@
 const fs = require("fs/promises");
 const { bootstrap } = require("kaholo-plugin-library");
-const { tryCreateRegexFromString, parsePath, pathExists } = require("./helpers");
+const {
+  tryCreateRegexFromString,
+  parsePath,
+  pathExists,
+} = require("./helpers");
 
 async function createFile({
   PATH: path,
   CONTENT: content,
+  overwrite,
 }) {
   const parsedPath = parsePath(path);
+  const fileAlreadyExists = await pathExists(parsedPath);
+
+  if (!overwrite && fileAlreadyExists) {
+    throw new Error(`File at ${path} already exists!`);
+  }
 
   try {
     await fs.writeFile(parsedPath, content);
@@ -14,7 +24,7 @@ async function createFile({
     throw new Error(`Failed to write to a file at ${path}: ${error.message || JSON.stringify(error)}`);
   }
 
-  return `Created file ${path}.`;
+  return fileAlreadyExists ? `File ${path} overwritten.` : `Created file ${path}.`;
 }
 
 async function appendToFile({
@@ -34,7 +44,7 @@ async function appendToFile({
   }
 
   if (!passedPathExists && doCreateFile) {
-    fileOperationResult = await createFile({ PATH: path, CONTENT: content });
+    fileOperationResult = await createFile({ PATH: path, CONTENT: content, overwrite: true });
   } else {
     try {
       await fs.appendFile(parsedPath, content);
@@ -74,7 +84,11 @@ async function replaceText({
   const fileContent = await getFileContent({ PATH: path });
 
   const newContent = fileContent.replace(parsedRegex, replaceValue);
-  await createFile({ PATH: path, CONTENT: newContent });
+  await createFile({
+    PATH: path,
+    CONTENT: newContent,
+    overwrite: true,
+  });
 
   return returnContent ? newContent : `File ${path} updated.`;
 }
